did useCurrentPath
